Added comment and additional ingress
Added middleware and ingress are to get around a limitation of custom DKP dashboards that do not allow anything other that an endpoint starting with a "/" for `dashboardLink`. This currently does not work for weave-gitops dashboard as it does not allow changing the path it is served on. It has hardcoded paths in it's javascript that does not work if endpoint is changed. Even middleware plugins like `rewritebody` don't work.